move print to dedicate file
